+++ cli/CMakeLists.txt
@@ -3,7 +3,9 @@ include_directories(inputhandler)
 
 add_subdirectory(inputhandler)
 
-add_executable(main main.cpp)
+set(HEADER_FILES commandhandler.h)
+
+add_executable(main main.cpp ${HEADER_FILES})
 
 target_link_libraries(main PUBLIC inputhandler)
 target_link_libraries(main PUBLIC utils)
